chore(refactor): vite - extract into buildRouteHooks.ts
