+++ packages/vite/src/buildFeServer.ts
@@ -1,15 +1,9 @@
-import type { PluginBuild } from 'esbuild'
-import { build as esbuildBuild } from 'esbuild'
 import { build as viteBuild } from 'vite'
 
-import {
-  getRouteHookBabelPlugins,
-  transformWithBabel,
-} from '@redwoodjs/babel-config'
 import { buildWeb } from '@redwoodjs/internal/dist/build/web'
-import { findRouteHooksSrc } from '@redwoodjs/internal/dist/files'
 import { getConfig, getPaths } from '@redwoodjs/project-config'
 
+import { buildRouteHooks } from './buildRouteHooks'
 import { buildRouteManifest } from './buildRouteManifest'
 import { buildRscFeServer } from './buildRscFeServer'
 import { ensureProcessDirWeb } from './utils'
@@ -82,43 +76,7 @@ export const buildFeServer = async ({ verbose, webDir }: BuildOptions = {}) => {
     logLevel: verbose ? 'info' : 'warn',
   })
 
-  const allRouteHooks = findRouteHooksSrc()
-
-  const runRwBabelTransformsPlugin = {
-    name: 'rw-esbuild-babel-transform',
-    setup(build: PluginBuild) {
-      build.onLoad({ filter: /\.(js|ts|tsx|jsx)$/ }, async (args) => {
-        const transformedCode = await transformWithBabel(args.path, [
-          ...getRouteHookBabelPlugins(),
-        ])
-
-        if (transformedCode?.code) {
-          return {
-            contents: transformedCode.code,
-            loader: 'js',
-          }
-        }
-
-        throw new Error(`Could not transform file: ${args.path}`)
-      })
-    },
-  }
-
-  await esbuildBuild({
-    absWorkingDir: getPaths().web.base,
-    entryPoints: allRouteHooks,
-    platform: 'node',
-    target: 'node16',
-    // @MARK Disable splitting and esm, because Redwood web modules don't support esm yet
-    // outExtension: { '.js': '.mjs' },
-    // format: 'esm',
-    // splitting: true,
-    bundle: true,
-    plugins: [runRwBabelTransformsPlugin],
-    packages: 'external',
-    logLevel: verbose ? 'info' : 'error',
-    outdir: rwPaths.web.distRouteHooks,
-  })
+  await buildRouteHooks(verbose, rwPaths)
 
   // Write a route manifest
   await buildRouteManifest()
